requirements.txt
# Dépendances principales
pyairtable>=2.2

# LangChain écosystème – versions compatibles
langchain==0.3.10
langchain-community==0.3.10
langchain-anthropic==0.2.4

# SDK officiel Anthropic (version requise par langchain-anthropic)
anthropic>=0.30,<1.0

# Vector DB & IA
pinecone>=7.0
boto3>=1.34

# Utilitaires
python-dotenv>=1.0
streamlit>=1.28
pandas>=2.0
plotly>=5.17 